Move NetworkFormattable to the correct package
